requirements.txt
dj-database-url==0.4.0
Django==1.9.2
djangorestframework==3.3.2
gunicorn==19.4.5
wheel==0.29.0
whitenoise==2.0.6
psycopg2==2.6.1